feat: add OverlayActionExecutor event sink validation and enhance test cases with DummySocket
