+++ tests/adapters/test_action_executor.py
@@ -1,34 +1,12 @@
 from __future__ import annotations
 
-import json
 import socket
 from pathlib import Path
 
 from adapters.action_executor import OverlayActionExecutor
 from adapters.dcs.overlay.sender import DcsOverlaySender
 from core.types import Event
-
-
-class DummySocket:
-    def __init__(self) -> None:
-        self.sent: list[tuple[bytes, tuple[str, int]]] = []
-        self._timeout: float | None = None
-
-    def settimeout(self, timeout: float) -> None:
-        self._timeout = timeout
-
-    def gettimeout(self) -> float | None:
-        return self._timeout
-
-    def sendto(self, data: bytes, server) -> None:
-        self.sent.append((data, server))
-
-    def close(self) -> None:
-        return None
-
-
-def _decode_cmd(data: bytes) -> dict:
-    return json.loads(data.decode("utf-8"))
+from tests.adapters.socket_stubs import DummySocket, decode_overlay_command
 
 
 def test_executor_maps_target_and_sends_highlight_udp(monkeypatch) -> None:
@@ -51,7 +29,7 @@ def test_executor_maps_target_and_sends_highlight_udp(monkeypatch) -> None:
     )
 
     assert len(dummy.sent) == 1
-    cmd = _decode_cmd(dummy.sent[0][0])
+    cmd = decode_overlay_command(dummy.sent[0][0])
     assert cmd["action"] == "highlight"
     assert cmd["target"] == "pnt_375"
     assert len(report.executed) == 1
@@ -149,7 +127,7 @@ def test_executor_max_targets_only_applies_to_overlay_execution(monkeypatch) -> 
     )
 
     assert len(dummy.sent) == 1
-    cmd = _decode_cmd(dummy.sent[0][0])
+    cmd = decode_overlay_command(dummy.sent[0][0])
     assert cmd["target"] == "pnt_375"
     assert len(report.executed) == 1
     assert len(report.rejected) == 1
@@ -187,7 +165,7 @@ def test_executor_system_pulse_controls_clear_by_ttl(monkeypatch) -> None:
 
     assert len(report.executed) == 1
     assert len(dummy.sent) == 2
-    cmds = [_decode_cmd(item[0]) for item in dummy.sent]
+    cmds = [decode_overlay_command(item[0]) for item in dummy.sent]
     assert cmds[0]["action"] == "highlight"
     assert cmds[0]["target"] == "pnt_375"
     assert cmds[1]["action"] == "clear"
@@ -205,4 +183,4 @@ def test_executor_does_not_mutate_external_sender_event_sink(monkeypatch) -> Non
 
     assert len(report.executed) == 1
     assert sender.event_sink is None
-    assert events == []
+    assert any(evt.kind == "overlay_requested" for evt in events)
